pyproject.toml
[project]
name = "skillsible"
version = "0.1.0"
description = "Ansible-style skill management for agent CLIs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
skillsible = "skillsible.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
